Next js로 마이그레이션 하라
next js 를 사용하도록 마이그레이션합니다.
