+++ babel.config.js
@@ -1,8 +1,14 @@
 module.exports = {
   presets: [
-    '@babel/preset-env',
-    ['@babel/preset-react', { runtime: 'automatic' }],
-    '@babel/preset-typescript',
+    [
+      'next/babel',
+      {
+        "preset-react": {
+          runtime: 'automatic',
+          importSource: '@emotion/react',
+        },
+      },
+    ],
   ],
-  plugins: ["@babel/plugin-transform-runtime"],
-};
+  plugins: ['@emotion/babel-plugin'],
+}
