+++ package.json
@@ -4,8 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "webpack serve --mode=development --open",
-    "build": "webpack --mode=production",
+    "dev": "next dev -p 80",
+    "build": "next build",
+    "start": "next start",
     "test": "jest",
     "lint": "npx eslint . --ext .ts,.tsx,.js,.jsx"
   },
@@ -14,11 +15,7 @@
   "private": "true",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@babel/plugin-transform-runtime": "^7.17.0",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
+    "@emotion/babel-plugin": "^11.10.5",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.0.0",
     "@types/jest": "^27.4.1",
@@ -28,38 +25,17 @@
     "@types/redux-mock-store": "^1.0.3",
     "@typescript-eslint/eslint-plugin": "^5.17.0",
     "@typescript-eslint/parser": "^5.17.0",
-    "babel-loader": "^8.2.4",
-    "copy-webpack-plugin": "^11.0.0",
     "eslint": "^8.12.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-unused-imports": "^2.0.0",
     "gh-pages": "^4.0.0",
     "given2": "^2.1.7",
-    "html-webpack-plugin": "^5.5.0",
     "jest": "^27.5.1",
     "jest-localstorage-mock": "^2.4.21",
     "jest-plugin-context": "^2.9.0",
     "redux-mock-store": "^1.5.4",
-    "typescript": "^4.6.3",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.8.1",
-    "webpack-merge": "^5.8.0",
-    "webpack-pwa-manifest": "^4.3.0",
-    "workbox-background-sync": "^6.5.4",
-    "workbox-broadcast-update": "^6.5.4",
-    "workbox-cacheable-response": "^6.5.4",
-    "workbox-core": "^6.5.4",
-    "workbox-expiration": "^6.5.4",
-    "workbox-google-analytics": "^6.5.4",
-    "workbox-navigation-preload": "^6.5.4",
-    "workbox-precaching": "^6.5.4",
-    "workbox-range-requests": "^6.5.4",
-    "workbox-routing": "^6.5.4",
-    "workbox-strategies": "^6.5.4",
-    "workbox-streams": "^6.5.4",
-    "workbox-webpack-plugin": "^6.5.4"
+    "typescript": "^4.6.3"
   },
   "dependencies": {
     "@babel/runtime": "^7.17.9",
@@ -71,6 +47,7 @@
     "@types/lodash": "^4.14.182",
     "axios": "^0.27.2",
     "lodash": "^4.17.21",
+    "next": "^13.1.6",
     "re-resizable": "^6.9.9",
     "react": "^18.0.0",
     "react-device-detect": "^2.2.2",
