--- src/index.tsx
@@ -1,39 +1,0 @@
-import { createRoot } from "react-dom/client";
-
-import { Provider } from "react-redux";
-
-import { QueryClient, QueryClientProvider } from "react-query";
-
-import { ReactQueryDevtools } from "react-query/devtools";
-
-import App from "./App";
-
-import { setupStore } from "./store";
-
-import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
-
-const root = createRoot(document.getElementById("app") as Element);
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      staleTime: Infinity,
-      refetchOnMount: false,
-      refetchOnWindowFocus: false,
-      retry: false,
-    },
-  },
-});
-
-const store = setupStore();
-
-root.render(
-  <QueryClientProvider client={queryClient}>
-    <Provider store={store}>
-      <App />
-    </Provider>
-    <ReactQueryDevtools initialIsOpen={false} />
-  </QueryClientProvider>,
-);
-
-serviceWorkerRegistration.register();
