+++ styles/globals.css
@@ -1,0 +1,8 @@
+html, body, #app {
+  margin: 0;
+  height: 100%;
+}
+
+#__next {
+  height: 100%;
+}
