+++ tsconfig.json
@@ -16,7 +16,8 @@
     "noEmit": true,
     "moduleResolution": "node",
     "resolveJsonModule": true,
-    "isolatedModules": true
+    "isolatedModules": true,
+    "incremental": true
   },
   "include": [
     "**/*.ts",
@@ -24,7 +25,7 @@
     "jest.setup.ts",
     ".eslintrc.js",
     "babel.config.js",
-    "jest.config.js",
+    "jest.config.js"
   ],
   "exclude": [
     "node_modules"
